This is still very much in progress, especially those gamemains and gametemplates. Only was in class for about 15 minutes.

+++ src/Player.java
@@ -59,7 +59,7 @@ public class Player implements PlayerInterface
 		color=c;
 		
 	}
-	public int getSpeed() {
+	public double getSpeed() {
 		return speed_x;
 	}
